refactor : inputView Test, resultView Test를 따로 분리하다.
